Disable crash reporter since Microsoft refuses to fix this upstream.
Unfortunately a PR submitted by a third-party was rejected by Microsoft, which would change the default for crash reporting from true to false and move the telemetry configuration to ensure crash reporting was not being forced on, placed in the argv.json, and cause a crash when improperly set.

So, disable it entirely. No cookies for you, Visual Studio Code team.

This package update also updates us to 1.59.1.

Properly resolves T9875 since Microsoft won't.